backend/requirements.txt
fastapi
boto3
uvicorn
python-multipart
requests
openai
langchain
pinecone-client
python-dotenv
# snowflake-sqlalchemy
sqlalchemy==1.4.51
snowflake-sqlalchemy==1.5.1
pandas
pydantic